+++ src/PhpWord/Writer/EPub3/Part/Meta.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * This file is part of PHPWord - A pure PHP library for reading and writing
  * word processing documents.
@@ -15,12 +16,12 @@
  * @license     http://www.gnu.org/licenses/lgpl.txt LGPL version 3
  */
 
-namespace PhpOffice\PhpWord\Writer\ePub3\Part;
+namespace PhpOffice\PhpWord\Writer\EPub3\Part;
 
 /**
- * Class for ePub3 content part.
+ * Class for EPub3 metadata part.
  */
-class Content extends AbstractPart
+class Meta extends AbstractPart
 {
     /**
      * Write part content.
@@ -30,18 +31,12 @@ class Content extends AbstractPart
     public function write()
     {
         $content = '<?xml version="1.0" encoding="UTF-8"?>';
-        $content .= '<package xmlns="http://www.idpf.org/2007/opf" version="3.0">';
-        $content .= '<metadata xmlns:dc="http://purl.org/dc/elements/1.1/">';
-        $content .= '<dc:title>Sample ePub3 Document</dc:title>';
+        $content .= '<metadata xmlns="http://www.idpf.org/2007/opf">';
+        $content .= '<dc:title>Sample EPub3 Document</dc:title>';
         $content .= '<dc:language>en</dc:language>';
+        $content .= '<dc:identifier id="bookid">urn:uuid:12345</dc:identifier>';
+        $content .= '<meta property="dcterms:modified">2023-01-01T00:00:00Z</meta>';
         $content .= '</metadata>';
-        $content .= '<manifest>';
-        $content .= '<item id="content" href="content.xhtml" media-type="application/xhtml+xml"/>';
-        $content .= '</manifest>';
-        $content .= '<spine>';
-        $content .= '<itemref idref="content"/>';
-        $content .= '</spine>';
-        $content .= '</package>';
 
         return $content;
     }
